pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pr-review-agent"
version = "1.0.0"
description = "AI-powered Pull Request Review Agent"
authors = [
    {name = "PR Review Agent Team", email = "team@prreviewagent.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["code-review", "ai", "pull-request", "github", "gitlab", "bitbucket"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Version Control :: Git",
]

dependencies = [
    "requests>=2.31.0",
    "gitpython>=3.1.32",
    "google-generativeai>=0.3.0",
    "flask>=2.3.2",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pytest>=7.4.0",
    "rich>=13.4.2",
    "bandit>=1.7.5",
    "radon>=6.0.1",
    "flake8>=6.0.0",
    "pyflakes>=3.0.1",
    "colorlog>=6.7.0",
    "click>=8.1.3",
    "tenacity>=8.2.2"
]

[project.urls]
Homepage = "https://github.com/pr-review-agent/pr-review-agent"
Documentation = "https://github.com/pr-review-agent/pr-review-agent#readme"
Repository = "https://github.com/pr-review-agent/pr-review-agent"
Issues = "https://github.com/pr-review-agent/pr-review-agent/issues"

[project.scripts]
pr-review = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/venv/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]